Add MilMove Modal component
